+++ classes/globals.php
@@ -29,66 +29,6 @@ class globals {
     public const AUTH = 'companion';
 
     /**
-     * Returns the \moodle_database instance
-     *
-     * @return \moodle_database
-     */
-    public static function db() {
-        global $DB;
-        return $DB;
-    }
-
-    /**
-     * Returns the \renderer_base instance
-     *
-     * @return \renderer_base
-     */
-    public static function output() {
-        global $OUTPUT;
-        return $OUTPUT;
-    }
-
-    /**
-     * Returns the $SESSION instance
-     *
-     * @return \stdClass
-     */
-    public static function session() {
-        global $SESSION;
-        return $SESSION;
-    }
-
-    /**
-     * Returns the $USER instance
-     *
-     * @return \stdClass
-     */
-    public static function user() {
-        global $USER;
-        return $USER;
-    }
-
-    /**
-     * Returns the $PAGE instance
-     *
-     * @return \moodle_page
-     */
-    public static function page() {
-        global $PAGE;
-        return $PAGE;
-    }
-
-    /**
-     * Returns the $CFG instance
-     *
-     * @return \stdClass
-     */
-    public static function cfg() {
-        global $CFG;
-        return $CFG;
-    }
-
-    /**
      * Returns the plugin configuration.
      *
      * @return \stdClass
@@ -103,33 +43,4 @@ class globals {
         return $mycfg;
     }
 
-    /**
-     * Returns the $SITE instance
-     *
-     * @return \stdClass
-     */
-    public static function site() {
-        global $SITE;
-        return $SITE;
-    }
-
-    /**
-     * Returns the $COURSE instance
-     *
-     * @return \stdClass
-     */
-    public static function course() {
-        global $COURSE;
-        return $COURSE;
-    }
-
-    /**
-     * Returns the $FULLME instance
-     *
-     * @return string
-     */
-    public static function fullme() {
-        global $FULLME;
-        return $FULLME;
-    }
 }