+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## Release notes
 
+### Version v4.2 (2023020200)
+* #1 remove abstraction class with moodle globals
+* #2 optimized privacy api and added privacy test
+* #3 fixed user_login() function
+* #4 companion login clears $_SESSION vars
+
 ### Version v4.1 (2022120700)
 * First feature ready version
 * Including phpunit test and behat